added authentication for endpoints , modified the logic based on auth

+++ server/src/services/cartService.ts
@@ -2,6 +2,7 @@ import httpStatus from 'http-status';
 import CartModel, { CartInterface } from '../models/Cart';
 import ApiError from '../utils/apiError';
 import { OrderStatusType } from '../types/commonTypes';
+import { IAuthRequest } from '../types/authTypes';
 
 const createCartService = async (CartBody: CartInterface) => {
   return CartModel.create(CartBody);
@@ -12,9 +13,13 @@ const getCartsService = async (
   options: {
     limit?: number;
     page?: number;
-  }
+  },
+  req:IAuthRequest
 ) => {
   const pageNum = (options.limit || 100) * (options.page || 0);
+  if(!req.user?.isAdmin){
+    filter = {...filter, user:req.user?._id}
+  }
   const order = await CartModel.find({ filter })
     .limit(options.limit || 100)
     .skip(pageNum)
@@ -22,8 +27,12 @@ const getCartsService = async (
   return order;
 };
 
-const getCartByIdService = async (id: string) => {
-  return CartModel.findById(id);
+const getCartByIdService = async (cartId: string, req:IAuthRequest) => {
+  const filter:any = {_id:cartId }
+  if(!req.user?.isAdmin) {
+    filter.user = req.user?._id
+  }
+  return CartModel.find(filter);
 };
 
 const updateCartByIdService = async (
@@ -32,19 +41,26 @@ const updateCartByIdService = async (
     product?: string;
     user?: OrderStatusType;
     quantity?:number;
-  }
+  },
+  req:IAuthRequest
 ) => {
-  const cart = await getCartByIdService(cartId);
-  if (!cart) {
+  const cart = await getCartByIdService(cartId, req);
+  if (!cart || cart.length !== 1) {
     throw new ApiError(httpStatus.NOT_FOUND, 'cart not found');
   }
-  Object.assign(cart, updateBody);
-  await cart.save();
-  return cart;
+  Object.assign(cart[0], updateBody);
+  await cart[0].save();
+  return cart[0];
 };
 
-const deleteCartByIdService = async (cartId: string) => {
-  const deletedCart = await CartModel.findByIdAndDelete(cartId);
+const deleteCartByIdService = async (cartId: string, req:IAuthRequest) => {
+  let deletedCart;
+  if(req.user?.isAdmin){
+    deletedCart = await CartModel.findByIdAndDelete(cartId);
+  }else{
+    deletedCart = await CartModel.find({_id:cartId, user:req.user?.id});
+    deletedCart = deletedCart.length > 0 ? deletedCart[0] : deletedCart
+  }
   if (!deletedCart) {
     throw new ApiError(httpStatus.NOT_FOUND, 'cart not found');
   }
